Add notification list
